IPMGs PROJECT/requirements.txt
streamlit
transformers
torch
pillow
opencv-contrib-python
sympy
numpy
